Can now serialize the entire gridmap
We are now able to serialize the entire grid map.
This took some effort and slight trickery but we did it. Its done.

Api was created in such a manner that a user will be force to create a serialized struct, serialization and deserialization methods for whatever visual property they make.

From there, our gridmap can serialze the map without care