Major changes
Revamped the serialization/deserialization process to be much more simpler. Now the user no longer has to declare extra methods, or create additional structs to hold their data.

Removed the use of Unity input system from Gridtester class so we dont have to ship the program with it.
Fixed a slight bug where when loading a saved map, we dd not account for the gridSize or chunkSize